split handshake code into its own files, so that Transport file is less massive
