+++ test/run.sh
@@ -2,7 +2,7 @@
 
 mkdir build 
 cd build
-cmake .. || exit 1
+cmake -DCMAKE_BUILD_TYPE=Debug .. || exit 1
 make || exit 1
 
 ./ixwebsocket_unittest ${TEST}
